+++ game-frontend/projects/game-frontend/src/app/components/reports-list/reports-list.component.ts
@@ -1,35 +1,49 @@
-import { Component, OnInit, ViewChild } from '@angular/core';
+import { Component, OnInit, ViewChild, OnDestroy } from '@angular/core';
 
-import { ModalComponent } from '@owge/core';
+import { ModalComponent, ScreenDimensionsService } from '@owge/core';
 
 import { ReportService } from '../../services/report.service';
 import { MissionReport } from '../../shared/types/mission-report.type';
 import { BaseComponent } from '../../base/base.component';
 
+interface NormalizedMissionReport extends MissionReport {
+  normalizedDate?: Date;
+}
+
 @Component({
   selector: 'app-reports-list',
   templateUrl: './reports-list.component.html',
-  styleUrls: ['./reports-list.component.less']
+  styleUrls: ['./reports-list.component.scss']
 })
-export class ReportsListComponent extends BaseComponent implements OnInit {
+export class ReportsListComponent extends BaseComponent implements OnInit, OnDestroy {
 
-  public reports: MissionReport[];
-  public selectedReport: MissionReport;
+  public reports: NormalizedMissionReport[];
+  public selectedReport: NormalizedMissionReport;
+  public isDesktop: boolean;
 
   @ViewChild('reportDetailsModal', { static: true })
   private _modal: ModalComponent;
 
   private _page = 1;
+  private _identifier: string;
 
   constructor(
-    private _reportService: ReportService
+    private _reportService: ReportService,
+    private _screenDimensionsService: ScreenDimensionsService
   ) {
     super();
+    this._identifier = _screenDimensionsService.generateIdentifier(this);
   }
 
   public async ngOnInit() {
     this.requireUser();
+    this._screenDimensionsService.hasMinWidth(767, this._identifier).subscribe(val => this.isDesktop = val);
     this.reports = await this._reportService.findReports(this._page).toPromise();
+    this.reports.forEach(report => report.normalizedDate = this._findReportDate(report));
+  }
+
+  public ngOnDestroy(): void {
+    this._screenDimensionsService.removeHandler(this._identifier);
   }
 
   public showReportDetails(report: MissionReport): void {
@@ -37,7 +51,7 @@ export class ReportsListComponent extends BaseComponent implements OnInit {
     this._modal.show();
   }
 
-  public findReportDate(report: MissionReport): Date {
+  private _findReportDate(report: MissionReport): Date {
     if (report) {
       return report.reportDate
         ? report.reportDate